desafio-1/pipeline/requirements.txt
pandas==2.3.3
psycopg2-binary==2.9.11
python-dotenv==1.1.1
SQLAlchemy==2.0.43
